refactor: MVC구조로 분리
