+++ src/main/java/lotto/util/BuyAmountValidator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 public class BuyAmountValidator {
     private BuyAmountValidator() {
