+++ src/test/java/lotto/model/domain/LottoMachineTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
