+++ src/test/java/lotto/util/LottoNumbersParserTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
